Enum sapnwrfc::binding::_RFC_RC

#[repr(u32)]

pub enum _RFC_RC {
    RFC_OK,
    RFC_COMMUNICATION_FAILURE,
    RFC_LOGON_FAILURE,
    RFC_ABAP_RUNTIME_FAILURE,
    RFC_ABAP_MESSAGE,
    RFC_ABAP_EXCEPTION,
    RFC_CLOSED,
    RFC_CANCELED,
    RFC_TIMEOUT,
    RFC_MEMORY_INSUFFICIENT,
    RFC_VERSION_MISMATCH,
    RFC_INVALID_PROTOCOL,
    RFC_SERIALIZATION_FAILURE,
    RFC_INVALID_HANDLE,
    RFC_RETRY,
    RFC_EXTERNAL_FAILURE,
    RFC_EXECUTED,
    RFC_NOT_FOUND,
    RFC_NOT_SUPPORTED,
    RFC_ILLEGAL_STATE,
    RFC_INVALID_PARAMETER,
    RFC_CODEPAGE_CONVERSION_FAILURE,
    RFC_CONVERSION_FAILURE,
    RFC_BUFFER_TOO_SMALL,
    RFC_TABLE_MOVE_BOF,
    RFC_TABLE_MOVE_EOF,
    RFC_START_SAPGUI_FAILURE,
    RFC_ABAP_CLASS_EXCEPTION,
    RFC_UNKNOWN_ERROR,
    RFC_AUTHORIZATION_FAILURE,
    _RFC_RC_max_value,
}

_RFC_RC

RFC return codes used by all functions that do not directly return a handle. Also used as error indicator in the structure #RFC_ERROR_INFO::code.

Variants
RFC_OKRFC_COMMUNICATION_FAILURERFC_LOGON_FAILURERFC_ABAP_RUNTIME_FAILURERFC_ABAP_MESSAGERFC_ABAP_EXCEPTIONRFC_CLOSEDRFC_CANCELEDRFC_TIMEOUTRFC_MEMORY_INSUFFICIENTRFC_VERSION_MISMATCHRFC_INVALID_PROTOCOLRFC_SERIALIZATION_FAILURERFC_INVALID_HANDLERFC_RETRYRFC_EXTERNAL_FAILURERFC_EXECUTEDRFC_NOT_FOUNDRFC_NOT_SUPPORTEDRFC_ILLEGAL_STATERFC_INVALID_PARAMETERRFC_CODEPAGE_CONVERSION_FAILURERFC_CONVERSION_FAILURERFC_BUFFER_TOO_SMALLRFC_TABLE_MOVE_BOFRFC_TABLE_MOVE_EOFRFC_START_SAPGUI_FAILURERFC_ABAP_CLASS_EXCEPTIONRFC_UNKNOWN_ERRORRFC_AUTHORIZATION_FAILURE_RFC_RC_max_value