Function sapnwrfc::binding::RfcCreateFunctionDesc

pub unsafe extern "C" fn RfcCreateFunctionDesc(
    name: *const SAP_UC, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_FUNCTION_DESC_HANDLE

Creates an empty function description with the given name.

Add parameter descriptions and exception descriptions to the new RFC_FUNCTION_DESC_HANDLE via RfcAddParameter() and RfcAddException().\n

After the handle was used for creating a container, any modifications are forbidden.

\in *name The name of the function module. \out *errorInfo Probably not enough memory left?! \return An empty function description that can be used for constructing hard-coded metadata.