Function sapnwrfc::binding::RfcCreateTransaction

pub unsafe extern "C" fn RfcCreateTransaction(
    rfcHandle: RFC_CONNECTION_HANDLE, 
    tid: *mut SAP_UC, 
    queueName: *const SAP_UC, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_TRANSACTION_HANDLE

Creates a container for executing a (multi-step) transactional call.

If queueName is NULL, tRFC will be used, otherwise qRFC. Use RfcInvokeInTransaction() to add one (or more) function modules to the transactional call. When sending this transactional call to the backend via RfcSubmitTransaction(), the backend will then treat all function modules in the RFC_TRANSACTION_HANDLE as one LUW.

\in rfcHandle Client connection to the backend, into which you want to send this tRFC/qRFC LUW. \in tid A unique 24 character ID. \in *queueName For tRFC set this to NULL, for qRFC specify the name of a qRFC inbound queue in the backend. \out *errorInfo Error information in case there is a problem with the connection. \return A data container that can be filled with several function modules.