Function sapnwrfc::binding::RfcGetCachedTypeDesc

pub unsafe extern "C" fn RfcGetCachedTypeDesc(
    repositoryID: *const SAP_UC, 
    typeName: *const SAP_UC, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_TYPE_DESC_HANDLE

Looks for a cached structure/table description.

Similar to RfcGetTypeDesc(), but it only looks into the cache. Again it should almost never be necessary to get a metadata description of a single structure.

\in *repositoryID System ID of R/3 System, for which this type description has been cached, or NULL in case you want to look for it in the default repository. \in *typeName Name of the DDIC structure/table. \out *errorInfo Requested type name not in cache? \return The requested type description.