Function sapnwrfc::binding::RfcGetMetadataQueryFailedEntry

pub unsafe extern "C" fn RfcGetMetadataQueryFailedEntry(
    handle: RFC_METADATA_QUERY_RESULT_HANDLE, 
    type_: RFC_METADATA_OBJ_TYPE, 
    index: c_uint, 
    entry: *mut RFC_METADATA_QUERY_RESULT_ENTRY, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Returns the error entry from the metadata query result

Returns the object name and the error text, if an error occurred during the query.

\in handle Handle to a metadata query result \in type Type of the requested object \in index Index of the requested object. Must be between 0 and "failed - 1", where "failed" is the value returned from RfcDescribeMetadataQueryResult() for the given RFC_METADATA_OBJ_TYPE. \inout *entry Pointer to the allocated structure to store the name of the metadata object and the error text \out *errorInfo More details in error case \return RFC_RC