Function sapnwrfc::binding::RfcGetPartnerSNCName

pub unsafe extern "C" fn RfcGetPartnerSNCName(
    rfcHandle: RFC_CONNECTION_HANDLE, 
    sncName: *mut SAP_UC, 
    length: c_uint, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Gets the partner's SNC name, if any.

\in rfcHandle RFC server connection. If this function is executed on a client connection, RFC_ILLEGAL_STATE will be returned. \out *sncName Pre-allocated buffer, which will receive the backend user's SNC name (null-terminated string). \in length Size of the pre-allocated buffer. This information is coming from the GSS library, therefore unfortunately the feature of assigning the used/required length to an output parameter is not possible in this case. The maximum length of an SNC name is 256. \out *errorInfo More error details in case SNC is not active. \return RFC_RC