Function sapnwrfc::binding::RfcInstallAuthorizationCheckHandler

pub unsafe extern "C" fn RfcInstallAuthorizationCheckHandler(
    onAuthorizationCheck: RFC_ON_AUTHORIZATION_CHECK, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Installs an optional callback function for performing authorization checks on incoming function calls.

After an RFC call arrives at the RFC library, and before the library executes the corresponding server implementation for that function module, it calls this callback function, if installed. Here you can implement a central authorization check, defining exactly which user from which backend system is allowed to execute what functions in your RFC server program. The actual function module implementations can then concentrate on their business logic and don't need to pay attention to access and authorization checks.

\in onAuthorizationCheck Pointer to a function of type RFC_ON_AUTHORIZATION_CHECK. The RFC lib calls this function, whenever an RFC request arrives from the backend. \out *errorInfo Additional information, in case the handler could not be installed. \return RFC_RC