Function sapnwrfc::binding::RfcInstallPassportManager

pub unsafe extern "C" fn RfcInstallPassportManager(
    onClientCallStart: RFC_PM_CALLBACK, 
    onClientCallEnd: RFC_PM_CALLBACK, 
    onServerCallStart: RFC_PM_CALLBACK, 
    onServerCallEnd: RFC_PM_CALLBACK, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Installs the necessary callback functions for processing Extended Passport (EPP) events.

Extended Passports is a feature that supersedes Distributed Statistical Records (DSRs). It allows to track certain performance attributes across system boundaries. These callback functions allow the application to exchange passports with the backend system. The actual data contained in the passports, can be processed with the eppslib library. Note: Normal RFC applications should never need to deal with this topic at all...

\in onClientCallStart Pointer to a function of type RFC_PM_CALLBACK. The RFC lib calls this function, before it sends a client request into the backend. Here you can provide the passport that is to be sent along with the RFC request. \in onClientCallEnd Pointer to a function of type RFC_PM_CALLBACK. The RFC lib calls this function, after the RFC response was received from the backend. You can read the updated passport information. \in onServerCallStart Pointer to a function of type RFC_PM_CALLBACK. The RFC lib calls this function, when it received an RFC request from the backend. Here you can read the passport that accompanied the request. \in onServerCallEnd Pointer to a function of type RFC_PM_CALLBACK. The RFC lib calls this function, before it will send the response to the above request back to the backend. You can update the passport data with information about processing performance of the current call. \out *errorInfo Additional information, in case the handlers could not be installed. \return RFC_RC