Function sapnwrfc::binding::RfcInstallPasswordChangeHandler

pub unsafe extern "C" fn RfcInstallPasswordChangeHandler(
    onPasswordChange: RFC_ON_PASSWORD_CHANGE, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Installs an optional callback function for processing password change events.

Whenever a logon attempt is made with a user, whose password is still initial, the backend system challenges this user to change his/her password. If this happens during a call to RfcOpenConnection(), the RFC library calls this callback function (if installed), and allows you to change the user's password. In your implementation, you can either return RFC_EXTERNAL_FAILURE, if you can't/don't want to change the password. In that case RfcOpenConnection() will end with RFC_LOGON_FAILURE and any additional text from errorInfo->message will be returned to the application. Or you can fill the old and new password with correct values and return RFC_OK, upon which the RFC library will attempt to change the password accordingly.

\in onPasswordChange Pointer to a function of type RFC_ON_PASSWORD_CHANGE. The RFC lib calls this function, whenever an initial password is detected during a call to RfcOpenConnection(). \out *errorInfo Additional information, in case the handler could not be installed. \return RFC_RC