Function sapnwrfc::binding::RfcInvoke

pub unsafe extern "C" fn RfcInvoke(
    rfcHandle: RFC_CONNECTION_HANDLE, 
    funcHandle: RFC_FUNCTION_HANDLE, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Executes a function module in the backend system.

The return codes have the following meaning: - RFC_OK\n The function call was executed successfully. - RFC_ABAP_EXCEPTION\n The function call was executed and ended with a defined ABAP Exception. The key of the exception can be obtained from errorInfo->key.

In the above two cases "rfcHandle" is still open and can be used to execute further function call.

RFC_ABAP_MESSAGE\n The function call was started to be processed, but was aborted with an ABAP Message. The message parameters can be obtained from errorInfo->abapMsgClass, errorInfo->abapMsgType, errorInfo->abapMsgNumber, errorInfo->abapMsgV1, ..., errorInfo->abapMsgV4.

RFC_ABAP_RUNTIME_FAILURE\n The function call was started to be processed, but was aborted with a SYSTEM_FAILURE (e.g division by zero, unhandled exception, etc in the backend system). Details can be obtained from errorInfo->message.

RFC_COMMUNICATION_FAILURE\n The connection broke down while processing the function call. Details can be obtained from errorInfo->message.

In these three cases the connection has been closed, so the "rfcHandle" needs to be refreshed via RfcOpenConnection.

RFC_INVALID_HANDLE\n "rfcHandle" is invalid or points to a connection that has already been closed.

\in rfcHandle Client connection over which to execute the function module. \inout funcHandle Data container containing the input data for the function module. %RfcInvoke() will write the FM's output data into this container. \out *errorInfo Additional error information. \return RFC_RC