Function sapnwrfc::binding::RfcIsAbapClassExceptionEnabled

pub unsafe extern "C" fn RfcIsAbapClassExceptionEnabled(
    funcHandle: RFC_FUNCTION_HANDLE, 
    isEnabled: *mut c_int, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Checks whether this function handle has been enabled for ABAP class exception support.

R/3 Systems from kernel release 7.11 on support ABAP class exception. Typically, an RFC server can use this function in order to find out, whether the function module currently being processed has ABAP class exceptions enabled in the ABAP backend. Only if this is the case, can the RFC Server safely throw an ABAP class exception.

\in funcHandle Function module data container. \out *isEnabled Flag indicating whether the function handle has been enabled for ABAP class based exception support. \out *errorInfo Should always return successfully. \return RFC_RC