Function sapnwrfc::binding::RfcMetadataBatchQuery

pub unsafe extern "C" fn RfcMetadataBatchQuery(
    rfcHandle: RFC_CONNECTION_HANDLE, 
    functionNames: *mut *const SAP_UC, 
    functionCount: c_uint, 
    typeNames: *mut *const SAP_UC, 
    typeCount: c_uint, 
    classNames: *mut *const SAP_UC, 
    classCount: c_uint, 
    handle: RFC_METADATA_QUERY_RESULT_HANDLE, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Queries the meta data for function, type and class lists.

Queries the meta data for function, type and class lists. All meta data is fetched using one roundtrip, the result is cached in the repository cache.

Note: You can use this feature only, if your backend system supports it. SAP Note 1456826 describes the required minimum support package level that the backend system must have. If the backend system supports it, open the RFC_CONNECTION_HANDLE with the additional logon parameter USE_REPOSITORY_ROUNDTRIP_OPTIMIZATION=1, before you pass it to %RfcMetadataBatchQuery().

\in rfcHandle Open client connection to the R/3 System, for which you need the function module, type or class descriptions. \in *functionNames Names of the function modules to look up \in functionCount Length of the function name list \in *typeNames Names of the types (structures and tables) to look up \in typeCount Length of the type name list \in *classNames Names of the classes to look up \in classCount Length of the class name list \in handle Query result, may be null \out *errorInfo More error details in case something goes wrong \return RFC_RC