Function sapnwrfc::binding::RfcRegisterServer

pub unsafe extern "C" fn RfcRegisterServer(
    connectionParams: *const RFC_CONNECTION_PARAMETER, 
    paramCount: c_uint, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_CONNECTION_HANDLE

Registers a server connection at an SAP gateway.

The connectionParams may contain the following name-value pairs: - gwhost, gwserv, program_id, trace, and the parameters for SNC communication as in RfcOpenConnection().

Program_id corresponds to an RFC destination in SM59 of type "T" in registration mode.

For a complete list of logon parameters to be used in connectionParams as well as in the sapnwrfc.ini file, see the sample sapnwrfc.ini file in the SDK's demo folder.

If the connection registration was ok, %RfcRegisterServer() returns a server connection handle, which can be used in RfcListenAndDispatch(). Otherwise the return value is NULL and errorInfo contains information similar to the RfcOpenConnection() case.

\in *connectionParams An array of RFC_CONNECTION_PARAMETERs with the names as described above and the values as necessary in your landscape. \in paramCount Number of parameters in the above array. \out *errorInfo Returns more error details, if the connect attempt fails. \return A handle to an RFC server connection that can be used for listening for function module requests from the backend.