Function sapnwrfc::binding::RfcSetString

pub unsafe extern "C" fn RfcSetString(
    dataHandle: DATA_CONTAINER_HANDLE, 
    name: *const SAP_UC, 
    stringValue: *const SAP_UC, 
    valueLength: c_uint, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Sets the given string value (stringValue/valueLength) into the field.

The field specified by name needs to be of one of the following data types: - RFCTYPE_CHAR - RFCTYPE_STRING - RFCTYPE_NUM - RFCTYPE_DATE - RFCTYPE_TIME - RFCTYPE_INTx - RFCTYPE_FLOAT - RFCTYPE_BCD - RFCTYPE_DECFxx - RFCTYPE_BYTE - RFCTYPE_XSTRING Note: If the target field is a numerical type, the RFC library tries to convert the string to a number. If the target field has type BYTE or XSTRING, the char value will be interpreted as hex encoded string representation of the bytes. Its length needs to be even in that case.\n Example: the 8 characters "CAFEBABE" will be converted to 4 bytes 0xCA 0xFE 0xBA 0xBE.

If the value cannot be converted to the desired target type, RFC_CONVERSION_ERROR will be returned.

\in dataHandle A data container (function handle, structure handle or table handle). If dataHandle is a table handle, the function will set the field value of the current row. \in *name The name of the field to set. \in *stringValue The characters to write into the field. \in valueLength The number of characters to read from stringValue. \out *errorInfo More information in case the field does not exist or a conversion fails. \return RFC_RC