Function sapnwrfc::binding::RfcSetTable

pub unsafe extern "C" fn RfcSetTable(
    dataHandle: DATA_CONTAINER_HANDLE, 
    name: *const SAP_UC, 
    value: RFC_TABLE_HANDLE, 
    errorInfo: *mut RFC_ERROR_INFO
) -> RFC_RC

Copies the given table into the target table of the parent container.

The target field needs to be of type RFCTYPE_TABLE.

Note: If you want to avoid the copy operation, use RfcGetTable() and set the subfields here, instead of the sequence RfcCreateTable()/ setting the subfields/ #RfcSetTable().

\in dataHandle A data container (function handle, structure handle or table handle). If dataHandle is a table handle, the function will set the field value of the current row. \in *name The name of the field to set. \in value The table to copy. \out *errorInfo Field does not exist or is not of type RFCTYPE_TABLE? \return RFC_RC