Struct sapnwrfc::binding::_RFC_ATTRIBUTES

#[repr(C)]

pub struct _RFC_ATTRIBUTES {
    pub dest: [SAP_UC; 65],
    pub host: [SAP_UC; 101],
    pub partnerHost: [SAP_UC; 101],
    pub sysNumber: [SAP_UC; 3],
    pub sysId: [SAP_UC; 9],
    pub client: [SAP_UC; 4],
    pub user: [SAP_UC; 13],
    pub language: [SAP_UC; 3],
    pub trace: [SAP_UC; 2],
    pub isoLanguage: [SAP_UC; 3],
    pub codepage: [SAP_UC; 5],
    pub partnerCodepage: [SAP_UC; 5],
    pub rfcRole: [SAP_UC; 2],
    pub type_: [SAP_UC; 2],
    pub partnerType: [SAP_UC; 2],
    pub rel: [SAP_UC; 5],
    pub partnerRel: [SAP_UC; 5],
    pub kernelRel: [SAP_UC; 5],
    pub cpicConvId: [SAP_UC; 9],
    pub progName: [SAP_UC; 129],
    pub partnerBytesPerChar: [SAP_UC; 2],
    pub partnerSystemCodepage: [SAP_UC; 5],
    pub reserved: [SAP_UC; 79],
}

_RFC_ATTRIBUTES

Structure returned by RfcGetConnectionAttributes() giving some information about the partner system on the other side of this RFC connection.

Fields
dest: [SAP_UC; 65]

RFC destination

host: [SAP_UC; 101]

Own host name

partnerHost: [SAP_UC; 101]

Partner host name

sysNumber: [SAP_UC; 3]

R/3 system number

sysId: [SAP_UC; 9]

R/3 system ID

client: [SAP_UC; 4]

Client ("Mandant")

user: [SAP_UC; 13]

User

language: [SAP_UC; 3]

Language

trace: [SAP_UC; 2]

Trace level (0-3)

isoLanguage: [SAP_UC; 3]

2-byte ISO-Language

codepage: [SAP_UC; 5]

Own code page

partnerCodepage: [SAP_UC; 5]

Partner code page

rfcRole: [SAP_UC; 2]

C/S: RFC Client / RFC Server

type_: [SAP_UC; 2]

2/3/E/R: R/2,R/3,Ext,Reg.Ext

partnerType: [SAP_UC; 2]

2/3/E/R: R/2,R/3,Ext,Reg.Ext

rel: [SAP_UC; 5]

My system release

partnerRel: [SAP_UC; 5]

Partner system release

kernelRel: [SAP_UC; 5]

Partner kernel release

cpicConvId: [SAP_UC; 9]

CPI-C Conversation ID

progName: [SAP_UC; 129]

Name of the calling APAB program (report, module pool)

partnerBytesPerChar: [SAP_UC; 2]

Number of bytes per character in the backend's current codepage. Note this is different from the semantics of the PCS parameter.

partnerSystemCodepage: [SAP_UC; 5]

Partner system code page

reserved: [SAP_UC; 79]

Reserved for later use