Struct sapnwrfc::binding::_RFC_ERROR_INFO

#[repr(C)]

pub struct _RFC_ERROR_INFO {
    pub code: RFC_RC,
    pub group: RFC_ERROR_GROUP,
    pub key: [SAP_UC; 128],
    pub message: [SAP_UC; 512],
    pub abapMsgClass: [SAP_UC; 21],
    pub abapMsgType: [SAP_UC; 2],
    pub abapMsgNumber: [RFC_NUM; 4],
    pub abapMsgV1: [SAP_UC; 51],
    pub abapMsgV2: [SAP_UC; 51],
    pub abapMsgV3: [SAP_UC; 51],
    pub abapMsgV4: [SAP_UC; 51],
}

_RFC_ERROR_INFO

Used in all functions of the NW RFC library to return detailed information about an error that has just occurred. This can be an error that the communication partner sent back to us, an error that occurred in the network layer or operating system, an internal error in the NW RFC library or an error that the application programmer (i.e. you) has committed...

Within a server function implementation, the application programmer (you) can return this structure to the RFC library in order to specify the error type & message that you want to send back to the backend.

Fields
code: RFC_RC

Error code. Should be the same as the API returns if the API has return type RFC_RC

group: RFC_ERROR_GROUP

Error group

key: [SAP_UC; 128]

Error key

message: [SAP_UC; 512]

Error message

abapMsgClass: [SAP_UC; 21]

ABAP message ID , or class

abapMsgType: [SAP_UC; 2]

ABAP message type, e.g. 'E', 'A' or 'X'

abapMsgNumber: [RFC_NUM; 4]

ABAP message number

abapMsgV1: [SAP_UC; 51]

ABAP message details field 1, corresponds to SY-MSGV1

abapMsgV2: [SAP_UC; 51]

ABAP message details field 2, corresponds to SY-MSGV2

abapMsgV3: [SAP_UC; 51]

ABAP message details field 3, corresponds to SY-MSGV3

abapMsgV4: [SAP_UC; 51]

ABAP message details field 4, corresponds to SY-MSGV4