Struct sapnwrfc::binding::_RFC_PARAMETER_DESC

#[repr(C)]

pub struct _RFC_PARAMETER_DESC {
    pub name: RFC_ABAP_NAME,
    pub type_: RFCTYPE,
    pub direction: RFC_DIRECTION,
    pub nucLength: c_uint,
    pub ucLength: c_uint,
    pub decimals: c_uint,
    pub typeDescHandle: RFC_TYPE_DESC_HANDLE,
    pub defaultValue: RFC_PARAMETER_DEFVALUE,
    pub parameterText: RFC_PARAMETER_TEXT,
    pub optional: RFC_BYTE,
    pub extendedDescription: *mut c_void,
}

_RFC_PARAMETER_DESC

Structure for reading (RfcGetParameterDescByIndex() or RfcGetParameterDescByName()) or defining (RfcAddParameter()) the properties of a parameter in a function module.

Fields
name: RFC_ABAP_NAME

Parameter name, null-terminated string

type_: RFCTYPE

Parameter data type

direction: RFC_DIRECTION

Specifies whether the parameter is an input, output or bi-directional parameter

nucLength: c_uint

Parameter length in bytes in a 1-byte-per-SAP_CHAR system

ucLength: c_uint

Parameter length in bytes in a 2-byte-per-SAP_CHAR system

decimals: c_uint

Gives the number of decimals in case or a packed number (BCD)

typeDescHandle: RFC_TYPE_DESC_HANDLE

Handle to the structure definition in case this parameter is a structure or table

defaultValue: RFC_PARAMETER_DEFVALUE

Default value as defined in SE37

parameterText: RFC_PARAMETER_TEXT

Description text of the parameter as defined in SE37. Null-terminated string.

optional: RFC_BYTE

Specifies whether this parameter is defined as optional in SE37. 1 is optional, 0 non-optional

extendedDescription: *mut c_void

This field can be used by the application programmer (i.e. you) to store arbitrary extra information.