Struct sapnwrfc::binding::_RFC_SERVER_CONTEXT

#[repr(C)]

pub struct _RFC_SERVER_CONTEXT {
    pub type_: RFC_CALL_TYPE,
    pub tid: RFC_TID,
    pub unitIdentifier: *mut RFC_UNIT_IDENTIFIER,
    pub unitAttributes: *mut RFC_UNIT_ATTRIBUTES,
}

_RFC_SERVER_CONTEXT

Used in RfcGetServerContext() for obtaining more information about the current incoming function call.

Fields
type_: RFC_CALL_TYPE

Specifies the type of function call. Depending on the value of this field, some of the other fields of this struct may be filled.

tid: RFC_TID

If type is RFC_TRANSACTIONAL or RFC_QUEUED, this field is filled with the 24 digit TID of the tRFC/qRFC unit.

unitIdentifier: *mut RFC_UNIT_IDENTIFIER

If type is RFC_BACKGROUND_UNIT, this pointer is set to the unit identifier of the LUW. Note: the pointer is valid only during the execution context of your server function.

unitAttributes: *mut RFC_UNIT_ATTRIBUTES

If type is RFC_BACKGROUND_UNIT, this pointer is set to the unit attributes of the LUW. Note: the pointer is valid only during the execution context of your server function.