Struct sapnwrfc::binding::_RFC_UNIT_ATTRIBUTES

#[repr(C)]

pub struct _RFC_UNIT_ATTRIBUTES {
    pub kernelTrace: c_short,
    pub satTrace: c_short,
    pub unitHistory: c_short,
    pub lock: c_short,
    pub noCommitCheck: c_short,
    pub user: [SAP_UC; 13],
    pub client: [SAP_UC; 4],
    pub tCode: [SAP_UC; 21],
    pub program: [SAP_UC; 41],
    pub hostname: [SAP_UC; 41],
    pub sendingDate: RFC_DATE,
    pub sendingTime: RFC_TIME,
}

_RFC_UNIT_ATTRIBUTES

If the external program is the sender of the bgRFC unit, this structure is used to set a bunch of special attributes that determine, how a bgRFC Unit will be processed in the backend. The fields user, client, tCode and program are optional. If left empty, the NW RFC lib will fill them with default values. The fields hostname, sendingDate and sendingTime should not be filled, the lib fills them, when the unit is submitted.

If the external program is the receiver of the bgRFC unit, you can use RfcGetServerContext() to obtain the values that were sent by the backend.

Fields
kernelTrace: c_short

If != 0, the backend will write kernel traces, while executing this unit.

satTrace: c_short

If != 0, the backend will write statistic records, while executing this unit.

unitHistory: c_short

If != 0, the backend will keep a "history" for this unit.

lock: c_short

Used only for type Q: If != 0, the unit will be written to the queue, but not processed. The unit can then be started manually in the ABAP debugger.

noCommitCheck: c_short

Per default the backend will check during execution of a unit, whether one of the unit's function modules triggers an explicit or implicit COMMIT WORK. In this case the unit is aborted with an error, because the transactional integrity of this unit cannot be guaranteed. By setting "noCommitCheck" to true (!=0), this behavior can be suppressed, meaning the unit will be executed anyway, even if one of it's function modules "misbehaves" and triggers a COMMIT WORK.

user: [SAP_UC; 13]

Sender User (optional). Default is current operating system User.

client: [SAP_UC; 4]

Sender Client ("Mandant") (optional). Default is "000".

tCode: [SAP_UC; 21]

Sender Transaction Code (optional). Default is "".

program: [SAP_UC; 41]

Sender Program (optional). Default is current executable name.

hostname: [SAP_UC; 41]

Sender hostname. Used only when the external program is server. In the client case the nwrfclib fills this automatically.

sendingDate: RFC_DATE

Sending date in UTC (GMT-0). Used only when the external program is server. In the client case the nwrfclib fills this automatically.

sendingTime: RFC_TIME

Sending time in UTC (GMT-0). Used only when the external program is server. In the client case the nwrfclib fills this automatically.